+++ packages/recurring-payments-job/src/services/govuk-pay-service.js
@@ -2,48 +2,8 @@ import { govUkPayApi } from '@defra-fish/connectors-lib'
 import db from 'debug'
 const debug = db('recurring-payments:gov.uk-pay-service')
 
-export const sendPayment = async preparedPayment => {
-  const createPayment = () => govUkPayApi.createPayment(preparedPayment, true)
-  const response = await createPayment()
-  if (!response.ok) {
-    throw new Error(`Unexpected response from GOV.UK Pay API. 
-      Status: ${response.status}, 
-      Response: ${JSON.stringify(await response.json())}
-      Transaction ID: ${preparedPayment.id}
-      Payload: ${JSON.stringify(preparedPayment)}
-    `)
-  }
-  return response.json()
-}
-
-export const getPaymentStatus = async paymentId => {
-  if (!paymentId) {
-    throw new Error('Invalid payment ID')
-  }
-
-  const fetchPaymentStatus = async () => {
-    try {
-      return await govUkPayApi.fetchPaymentStatus(paymentId, true)
-    } catch (e) {
-      console.error('Error fetching payment status', paymentId)
-      throw e
-    }
-  }
-
-  const response = await fetchPaymentStatus()
-
-  if (!response.ok) {
-    console.error({
-      method: 'GET',
-      status: response.status,
-      response: await response.json(),
-      paymentId
-    })
-    throw new Error('Unexpected response from GOV.UK Pay API')
-  }
-
-  return response.json()
-}
+export const queueRecurringPayment = (preparedPayment, batcher) => govUkPayApi.queueRecurringPayment(preparedPayment, batcher)
+export const queueRecurringPaymentStatusCheck = (paymentId, batcher) => govUkPayApi.queueRecurringPaymentStatusCheck(paymentId, batcher)
 
 export const isGovPayUp = async () => {
   const response = await govUkPayApi.isGovPayUp()